Use forked lspconfig to support angularls in Nx workspace ¯\_(ツ)_/¯
There were an issue with using angularls in a newer Nx workspace,
because it doesn't have angular.json or workspace.json files anymore. I
tried to edit my lspconfig in my nvim config, though it seemed like it
was ignored completely. So, I tried to edit it directly in the
lspconfig's source code, and after that it started to work suddenly.

I'm guessing though I have may have some issues in my nvim config that I
should resolve, and cleanup and maybe that could make it work as well.
For a quick fix it is good enough for now.